server/requirements.txt
fastapi
uvicorn
authlib
cryptography
httpx
jwt
pyotp
qrcode
pydantic_settings
sqlalchemy
passlib
pydantic[email]
email-validator
psycopg2
jinja2
python-multipart
pyjwt
bcrypt
python-jose[cryptography]
websockets
uvicorn[standard]
wsproto
requests